refactor(ui): extract TrainingView logic to custom hooks and CSS
- Move audio processing to useAudioProcessing hook
- Move state management to useTrainingSession hook
- Separate inline styles into TrainingView.css
- Keep TrainingView as a pure presentation component
